Refactored my code to where the UICollectionView delegate and datasource methods are in the tableview cell file and now it populates the fake data correctly

+++ AppStore/Models/Apps.swift
@@ -13,6 +13,6 @@ class App {
     var id: Int?
     var title: String?
     var category: String?
-    var price: Int?
+    var price: Double?
     var appImage: String?
 }
